requirements.txt
asgiref==3.7.2
Django==4.2.7
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
phonenumbers==8.13.26
Pillow==10.1.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
